requirements.txt
numpy
numpy-stl